fix(react): update handleSize function on the useBagItem function
This prevents the handleSize function from updating a size
and adding the same size after. This happens because when
the previous bag quantity is bigger than the merchant quantity
for the size change, the function handleAddOrUpdateBagItem
is called with all the merchants of the size, instead of being
called with all the merchants, except the one that is updated previously.

Also, this updates the react/headers/index file in
order to have some consistency in ours headers
folder across the multiple packages in this
repo.