fix(redux): make `getItemWholeQuantity` return 0 when there's no bag
The `getItemWholeQuantity` selector uses internally the `getBagItems` selector
and then iterate through the bag items to get the whole quantity. When the `getBagItems`
runs before the bag is fetched, its return value is `undefined` by default, which throws
an error when trying to `reduce`. This fix avoids that and makes this selector
return 0 if there are no bag or bag items.