fix(redux): make `getBagItems` return empty array
Our `getBagItems` selector is returning `undefined` when the bag or bagItems don't exist but that
output breaks consistency and might create error in other selectors that use `getBagItems` without
validating if it is an array or undefined. To validate if the bag is fetched or not we have other,
more specific, selectors.