feat(client|redux): add order activities endpoints
This feature implementation is tackling the Confirm Delivery capability.

Key objective can be explained by allowing non-delivered orders
to be acknowledge as received by the costumer.

With this, the return process can start right away, instead of having to wait
for external mechanisms to update the state (process that can take some time,
depending on the carrier).