fix: remove condition from `useProductsList`
The `useProductsList` hook was calling the fetch list action only if it wasn't fetched before, even
tho this logic might look correct, it isn't. We need to call `fetchListing/fetchSet` anyway because
these actions are responsible for checking if the list is fetched and, if that's the case, just
replace the active "hash" in the store. By not calling any action, the active "hash" was not being
updated and the selectors who use the default "hash" value to get info would return the wrong list's
value.