fix(analytics|react): fix typescript typings
- This fixes some typescript typings for the analytics
package and replaces jsdoc with tsdoc. The `docs`
script (and all dependencies used by it) were removed
as they would not work with tsdoc.
- Some introductory work was done for using project
references in typescript in order to support the `analytics`
package compilation that imports package.json which
would not work with a single ts configuration file because
the rootDir was `src` and package.json was outside it.
Using project references in typescript allow to have the
package.json included alongside the src folder while also
keeping the layout for the compiled package to not include
the `src` folder.