fix(core): add `groupsOn` to the sizes facet's entity id
This adds the `groupsOn` value to the facet entity id for sizes, which fixes a bug of
having wrong descriptions of sizes. Since each size facet can have different descriptions
according to what category it belongs to (hence the `groupsOn`, which has a category id as its
value), there is the need to add the `groupsOn` to turn the entity id unique, despite of the
facet id.

For example, the the entity `sizes_17` could have both the values `36` or `XS`,
depending on the listing the user is. This would result in a bug when navigating between cached
listings, where the previous value for a given entity was kept, displaying the wrong descriptions
(like keeping `XS` for shoes). With this change, the entity id is now `sizes_17_<groupsOn>`, just
like the "sizes by category" already does.