fix(blackout-core): fix doUpdateUserSubscriptions action
This fixes the `doUpdateUserSubscriptions` action in order to reflect
the change made by the backend that changed the result of the put method
used by the action to not return the new subscriptions data and
instead requiring that a new get user subscriptions request be
performed to retrieve the updated data.
Additionally a new state for the error message of the
doGetUpdateUserSubscriptions action was added in order to allow the
consumer to differentiate the case when there was an error updating the
subscriptions and present a different error message.
To provide backwards compatibility, the previous `error` state was
maintained as well.