fix(blackout-client|blackout-redux): fix updateUserSubscriptions action
- This fixes the `updateUserSubscriptions` action in order to reflect
the change made by the backend that changed the result of the put method
used by the action to not return the new subscriptions data and
instead requiring that a new get user subscriptions request be
performed to retrieve the updated data.
Additionally a breaking change was added with the
`getUserSubscriptionsError` selector that will not return an error when
the error was from the update subscriptions action.
To get the update error, a new selector was introduced.
This was in order to better differentiate which kind of error occurred
with the subscriptions state and enable the developer
to show alternative UIs in that case.
- Some fixes in typescript typings for `entities` and `subscriptions` slices.

BREAKING CHANGE: The `getUserSubscriptionsError` selector will
not return the error when there is an error
for the `updateUserSubscriptions`
action. To check if there was an error
for that action, you will need to use the
`getUpdateSubscriptionsError` selector as the following example shows:

```
// Previously

import {
    getUserSubscriptionsError,
} from '@farfetch/blackout-redux/subscriptions';

const mapStateToProps = state => {
    return {
      // errorLoadingUserSubscriptions would return all errors
      errorLoadingUserSubscriptions: getUserSubscriptionsError(state),
    };
};

// Change to

import {
    getUserSubscriptionsError,
    getUpdateSubscriptionsError,
} from '@farfetch/blackout-redux/subscriptions';

const mapStateToProps = state => {
    return {
      // Now getUserSubscriptionsError will return errors for loading / deleting actions
      errorLoadingUserSubscriptions: getUserSubscriptionsError(state),
      // getUpdateSubscriptionsError will return the error for the update action
      errorUpdatingUserSubscriptions: getUpdateSubscriptionsError(state)
   };
};
```