feat: add the `facetId` to the `filterSegments` on the `searchResult` entity
This adds the `facetId` to the `filterSegments` on the `searchResult` entity in order to allow a
direct mapping between a filter segment and a facet entity. This will enable, for example, using the
`getFacet` selector from `filterSegment` values.

This also fills the `description`, instead of receiving `null` for the facets that have no slug
(hydrated from the facet itself).