fix(blackout-client|blackout-redux): fix getFormSchemaByCode selector
- This fixes the getFormSchemaByCode selector which was adding a default
value of an array when the schema was not found which would be
incorrect as it should return undefined in that case.
- Improved typings for FormSchema to match the type for the
json-schema package. Note: @types/form-schema is defined as a dependency
because our typings depend on it and typescript requires that we add
this as a dependency so the user of this package can browse to its
definitions.
- Fixed type errors in forms reducer testing file.