feat: implement next version
- This commit implements all required changes for the release
of a proper next version of our packages which include:
  - Compilation of typescript runtime files which will remove the need
   for ts-loader configuration in the tenants projects.
  - Pipeline typescript checking: There is now a step in the pipeline
   which will check if there are any type errors. For now, it will
   only check the runtime files. Test files will be added later when
   all type errors are resolved.
  - Revamped names of actions/clients. There are many name changes
   of actions and clients (mostly) to make it more consistent and
   intuitive. This is not yet completely finished.
  - Imports can now be made directly from the root of each package
   or, in alternative, to a specific folder to fix some dependency
   problems in our packages. For example, if you want to use analytics
   modules from react package, redux will need to be installed if you import
   from the package's main index.js.
  - Completely revamped the `users` folder in redux to be more manageable.
  - Many other changes.

Many apologies for such a big commit but time was (very) short so it was easier
to have only one branch and commit to manage instead of having many branches and
managing all of them.

BREAKING CHANGE: Many renames of actions and clients, as well as redux store
layout were implemented and will be described in a new migration file which
will be authored later.