feat(core|react): add analytics SPA page referrer context
Since document.referrer stays the same on single page applications,
this PR adds an alternative property called `pageLocationReferrer`
for page trackings that will hold the previous URL when navigating trough
pages.
This PR also adds the type of event that is being tracked to each `useContext`
function and calls the webContext directly on the analyticsWeb constructor.
Mapped the Omnitracking's `referrer` to the new `pageLocationReferrer`,
since the document.referrer was causing issues when analyzing user behavior.