fix: fix user typings
- This commit unifies the types `GuestUser` and `GuestUserNormalized`
to make development easier. This distinction existed because
the `getGuestUser` and `postGuestUser` clients return a
different type than the one returned by the `getUser` endpoint when the
user is guest. With this change it will
be easier to interact with user values.
- Remove source maps and declaration maps option for
compilation as they are useless because we are not publishing the
source code of the package.
- Fix some uses of `any` for the right type.
- Fix some lint warnings.