chore(client|react): remove authentication interceptor
- This removes any related authentication interceptor
functionality from the client and react packages as that solution will
only apply for native apps. This functionality will be migrated
to a package inside the react-native repository.
- Changed User typings to be more easy to use by consumers.
- Fixed PostWishlistItemActionData type to remove lingering
props for analytics tracking.
- Fixed parameter type for createContinentsList function.

BREAKING CHANGE: Authentication interceptor is now removed from both
client and react packages as it only makes sense for native apps.