fix: change getCountries client
- This changes the getCountries client to return all countries
transparently to the consumer instead of returning a paginated
response from the service. In the future, the service will change to
return all countries instead of return a paginated response as well.

BREAKING CHANGE: The getCountries client now does not return a paginated
response and does not accept any query parameters. The result is
an array of all the countries available.