feat(react): add orders hooks
- This adds the `useOrder` and `useOrders` hooks to the `react` package.
- Fixed normalization of the fetch orders response `entries` field to remove duplicates
and changed the `totalItems` calculation to the total number
of items after merging of the duplicates.
- Removed `ordersList` state from orders reducer and
changed the implementation of the root reducers to reflect
the state of the fetch orders request and because of that, the selectors
`isOrdersListLoading` and `getOrdersListError` were removed.
- Add normalization logic to map legacy order data (which
comes from the getGuestOrderLegacy client) to the Order
data type that comes from the `getOrder` client to ease
usage by developers so they do not need to treat order
values differently.
- Added the `resetOrderDetailsState` action to enable the
reset of the `orderDetails` slice.
- Added fetched selectors for user orders and order details requests.
- Added denormalization code to orders selectors.
- Some order and returns related type fixes.
- Removed initialReduxState from redux package fixtures as it was
only being used in one test.

BREAKING CHANGE: The selectors `isOrdersListLoading` and
`getOrdersListError` were removed and replaced with the selectors
`areOrdersLoading` and `getOrdersError`.
The orders reducer was now changed as well and not it
will only change its root slice `isLoading` and `error` values only
when the fetchUserOrders and fetchGuestOrders actions are used.