chore(react|redux): remove checkout hooks
- This removes `useCheckout` hook temporarily in order to not introduce
a breaking change when the next version of the package is released.
The redux implementation of both checkout and payments
areas were checked and tweaked a little bit by fixing reducers
and adding missing reset state actions.
- Renamed `getPaymentMethods` and `fetchPaymentMethods` to
`getCheckoutOrderPaymentMethods` and `fetchCheckoutOrderPaymentMethods`
respectively and moved those modules to the checkout area from payments.
- Changed `renderContent` function signature to accept a boolean
parameter instead of a string.
- Changed reducerFactory/createReducerWithResult exports to remove the
`validateReset` function call
and allow to pass action types that should reset
the reducer's state.

BREAKING CHANGE:
- `useCheckout` hook was removed from @farfetch/blackout-react package.
- `getCheckoutOrderOperation` and `fetchCheckoutOrderOperation` signatures
have changed to accept separate parameters for the checkoutOrderId and
operationId values instead of using an object.
- `getPaymentMethods` client was renamed to `getCheckoutOrderPaymentMethods`.
- `fetchPaymentMethods` action was renamed to `fetchCheckoutOrderPaymentMethods`.