fix(react): fix add/update bag item errors in useBag hook
- This adds throw statements to `handleAddUpdateBagItem` and
`handleFullUpdate` functions used by the actions in `useBag` hook in
order to communicate to the consumer that it was unable to
perform the action. This replaces the early return strategy which
would not communicate to the user that the actions were unable
to perform an add/update of a bag item because there was either
missing information or no stock available.