chore: fix release step of the pipeline
- This fixes the release step of the pipeline by running `lerna version`
before running the build script in order to have the
package.json files updated so that the babel transformer will
pick the correct package version when transforming the
imports that read the version from package.json files.
Currently, new releases of the analytics package will have the
incorrect version inserted by the babel transformer inside the
`constants.ts` file which was being used in error/warning
messages, which would confuse the consumer of the
package as it would be displaying the incorrect version that was
installed.

Example:
Package version: `1.0.0-next.92`.
Version inserted in `constants.ts` by babel transformer:
`1.0.0-next.91`.