fix(react|redux): fix wishlist and bag selectors
- This fixes `getWishlistItems` selector that
was recalculating its value unnecessarily every time the state changed
even though it was using `createSelector` because of an incorrect
function input to createSelector that was selecting the whole state.
- Changed `getBagItemAvailableSizes` selector to not use `getBagItem`
in its implementation as it would invalidate the cached result
created by createSelector as it would change the bag item id parameter.
- Refactored `getBagItems`, `getBagItem`, `getWishlistItems`
and `getWishlistItem` selectors to use a common denormalize function.
- Changed `useWishlist` and `useBag` hooks to returned
undefined when the wishlist and bag are undefined respectively.
This is to avoid generating a new object reference when the
data has not changed.
- Changed `renderContent` function to accept only 2 arguments instead
of multiple arguments and made the 2nd argument optional to avoid type
errors when the components registered do not use the same props as the
ones provided by the package.
- Added a `.filter(Boolean)` call in getProductDenormalized selector
for the categories property calculation to avoid receiving an array with
undefined values.