fix: add consistency fixes
This adds a set of consistency fixes to some inconsistencies that were
detected throughout a few modules inside all packages and other fixes:

- Renamed `useOrders` and `useReturns` hooks to `useUserOrders`
and `useUserReturns` respectively to be consistent with the
names of other hooks that provide similar logic.
- Changed the order of the `returnId` parameter in actions that allow
overriding the hook's `returnId` parameter to the last position
as overriding the return id is not a common usage and being in
the last position makes the methods easier to use for the much more
common usages of not overriding the `returnId` from the hook. This was
changed in hooks `useReturnPickupCapabiliity`,
`useReturnPickupRescheduleRequest` and
`useReturnPickupRescheduleRequests`.
- Changed all hooks' auto fetch condition to use `isFetched` instead of
looking at the error.
- Changed all hooks that used custom logic to calculate if the current
user is authenticated or not to use the `isAuthenticated` user'
selector instead.
- Changed useBag's `fetch` action to implicitly use the bagId passed
when invoking the hook instead of requiring the consumer to
specify it again.
- Changed useBagItem implementation to use a new selector,
isBagItemFetched to derive the information if a bag item is fetched
or not to be consistent with other hooks.
- Changed useUser's `changePassword` and `update` actions to implicitly
use the `userId` that is set in the store instead of requiring the
consumer to specify it.
- Changed useWishlist's fetch action to implicitly use the
wishlistId that was passed to the hook instead of requiring the
consumer to specify it.
- Changed useWishlistItem's `update` and `remove` actions to implicitly
use the `wishlistItemId` that was passed to the hook
instead of requiring the consumer to specify it. Also added a
`isWishlistItemFetched` selector to derive the value of the
`isFetched` property.
- Changed the implementation of the hooks `useWishlistSets`
and `useWishlistSet` to be consistent with the remaining hooks.
- Changed the `bags` and `wishlist` reducers to reset its state
when a fetch user request is performed and returns
different bag or wishlist ids.
- Unified the function signatures of the putUserDefault{Billing,Shipping,
Contact}Address clients.
- Changed the imports that were used in @example tags inside tsdoc to
point to the root of the package (when necessary).
- Added a new order status to the OrderStatus enum.

BREAKING CHANGE:
- `useOrders` and `useReturns` were renamed to
`useUserOrders` and `useUserReturns` respectively.
- `useBag`, `useBagItem`, `useWishlist`, `useWishlistItem`, `useUser`,
`useWishlistSet`, `useWishlistSets` `useReturnPickupCapabiliity`,
`useReturnPickupRescheduleRequest` and
`useReturnPickupRescheduleRequests` have some changes in its
public properties/actions. Check the description for more information.
- `putUserDefaultShippingAddress` and `putUserDefaultBillingAddress`
signatures changed. Now, instead of receiving an object containing the
userId and id properties, there is a specific parameter for both of
those properties.