feat(analytics|react): add forter and riskified integrations
- Add Forter and Riskified integrations;
- Remove vitorino;
- Refactor of current tests and add new ones.

throws an error and added a unit test for it.
- Changed `skipLibCheck` to false to allow type checking of `.d.ts` files
in the project.
- Added `@types/jquery` package to allow typing calls to jquery.
- Removed packages/redux/tsconfig.ci.tests.json as it was not needed
anymore.
- Added `dist` folder exclusion to `tsconfig.ci.tests.json` since
now with `skipLibCheck` set to false, it would catch the files from
`dist` folders and would give errors if they were available from a
 previous compilation.

BREAKING CHANGE: Vitorino was removed from the list of available
integrations and now there are separate Riskified and Forter integrations
that will need to be added separately to have the same behaviour
provided by Vitorino.