fix(redux): fix product attributes server render
This changes the serverInitialState handling of the
`productAttributes` property from page model to discard it if the value
is an empty array since the server will return an empty array by default
even when no request to fetch the product attributes was made.
Without this change, it was impossible to distinguish if the product
has no attributes or if the attributes were not fetched and the client
should then fetch them.

Also, added a fix for the DeliveryBundle type in checkout which
was missing some properties.