fix(redux): fix checkout reducer
- This removes the handling of the LOGIN_SUCCESS, REGISTER_SUCCESS
and FETCH_USER_SUCCESS actions from the checkout reducer since there
is a valid use case where the user wants to register/login after
completing a checkout order and it is necessary to keep the checkout
order data in the store.

BREAKING CHANGE: Now when the default login, register and fetch user
actions are dispatched to the store, no cleanup of any checkout
order state is performed.