chore: update eslint configuration
This applies the following changes for linting configuration:

- Added `jest/recommended`, `jest/style` and `jest-formatting/recommended`
to the `extends` option. This will add a number of checks that will
improve the consistency of the code used in unit tests.
- Set `curly` rule to 'error' to prevent statements without blocks
like ifs, whiles, etc...
- Set `@typescript-eslint/consistent-type-imports` to error with a
`fixStyle` set to `inline-type-imports` so that we can still use the
`no-duplicate-imports` rule. This rule adds type annotations to imports
that are used as types that will be used to bundlers as a hint that
they can drop those imports. NOTE: Currently typescript does not permit
having type annotations in imports that import types from both the
default export and named exports of a module. The workaround was to
change the import of default exports by using an alias like this:
`import { type default as XXX }` and with that we can still have
the `no-duplicate-imports` rule.
- Replaced `padding-line-between-statements` for `@typescript-eslint/
padding-line-between-statements` rule and added more rules to have
code formatted with more space.
- Set `require-await` rule to error to detect unnecessary async
function declarations when they not contain an `await` instruction.
- Changed the `.eslintrc.json` file extension to js since it is more
flexible and added the missing `root` property to true.

An experiment was tried to also use the `recommended-requiring-type-checking`
config from the `@typescript-eslint` package but due to the huge amount
of errors which were not really critical, it was removed. But that detect
a few unnecessary type statements and found a bug with the `initialize`
method in `packages/analytics/src/User.ts` which was not awaiting the
call to `set` method.