fix(redux): fix useBag hook actions
This fixes the `useBag` hook actions which were using stale `items`
data when the actions were invoked in succession (like a call to
removeItem is awaited and then a call to addItem is made).
The fix is to remove the `items` from the callbacks and use `getState`
to get the most updated state from the store.
Also, it was needed to await the `dispatch` calls to the bag success
actions to make sure that when the `.then` callback is invoked all
the middlewares have run first and the state in the store was updated.
We will need to change all factory actions to await the dispatch of
the success action (at least) eventually but on this PR I only included
wishlist, bag and user authentication actions as they are the most
problematic.