fix: more consistency fixes
- Changed `setBagPromocodesFactory` implementation to not store data
in an entity as it was not necessary and stored the data in `bags`
state directly.
- Fixed the name of `bagMiddlewares` export to `bagsMiddlewares` to
keep consistency with other middlewares names.
- Fixed type of the `customerType` property in `CheckoutOrder` type to
be a number instead of a string enum and added
new property `promotionEvaluationId` to it.
- Removed all actions and modules related to `returnPickupRequests`
since they are not necessary as their state is not being stored in
redux.

BREAKING CHANGE: `bagMiddleware` export was renamed to
`bagsMiddleware`. Also all redux modules related to
`returnPickupRequests` were removed. You will need to use the
`useReturnPickupRescheduleRequests` and
`useReturnPickupRescheduleRequest` hooks instead.