chore: add changes from patch
This adds some changes that were made in a patch in the boilerplate
application:
- Changed facet schema to use the `key` property instead of the
`description` property to generate the id for the facet as the
`description` might contain spaces which are not suitable to use as ids.
- Changed lists `serverInitialState` to calculate an error state for
when no products are returned from a listing page view model.
- Improved `Designers` type to add missing properties and add the
possibility of them being `null`.