fix: remove `contentTypeCode` as required parameter
This removes the `contentTypeCode` parameter from the required
parameters of `getSearchContents` client and fetchContents action.
Also, the hook `useContentType` was renamed to `useContents` since the
`contentTypeCode` parameter was removed so it did not make sense to
keep the `useContentType` name.

BREAKING CHANGE: `useContentType` hook is now called `useContents`
and the `contentTypeCode` parameter was removed. You will need to pass
 the `contentTypeCode` property in the options parameter now.