chore: add `release:git` script
This adds a `release:git` script which which will allow developers to
install a version of the packages containing the changes they are
developing without releasing them to npm. This script aims to
facilitate QA process by using gitpkg underneath to publish a built
version of the packages to a git repo of their choosing. The repo will
receive a commit and a tag which can be used to add the dependency in
an application so it can be sent for QA validation.