chore: update typescript to v5
This updates typescript to v5.0.2 which deprecates the
`importsNotUsedAsValues` compiler option in favour of
`verbatimModuleSyntax` option. Also changed the file extensions of
`axiosErrorSerializer`, `setupReactTestingLibrary` and `setupAfterEnv`
files to `mts` to solve an error with the upgrade to typescript.
A change was also made to `AuthenticationTokenManager` to fix a type
error detected by the upgraded version.