chore: change fetchCollectPoints action
This changes the fetchCollectPoints action to allow the query
parameter to be optional as its client specifies. To do this, the
layout of the checkout redux slice has changed to index the state
of those requests by query. The selectors now need to specify the
query parameter that matches the request done via `fetchCollectPoints`
to get the collect points request state.

Updated also the `collectCoverageFrom` value in jest.config to remove
unnecessary folders.

BREAKING CHANGE: The `fetchCollectPoints` action now does not
require the `query` parameter.
The following selectors to obtain state of collect points requests now requires
the query
parameter to match the query of the request:
- getCollectPoints
- areCollectPointsLoading
- getCollectPointsError
- areCollectPointsFetched

- The checkoutOrder returned by `getCheckoutOrder` selector now does
not contain the collectPoints property.