chore: prevent automatic major version bumps to `main`
This changes the pipeline to prevent commits that contain
breaking changes from updating the major version of the packages
as those are only supposed to be created when next is merged into main.

Unfortunately, the only was to make this work with lerna is to apply
a patch to the version command to look for a specific flag parameter
that was added for it.

This also applies some best practices to load github context variables
into run scripts in CI workflow.