chore: add support for legacy user endpoints
This adds support for the legacy user endpoints `getUserLegacy` and
`postUserLegacy` and its respective actions to avoid having patches in
applications where `getUser` and `fetchUser` do not work.

`useUser` hook now has a new option `useLegacyActions` which will call
the appropriate legacy actions for the `fetch` and `register` actions
if true.