fix: fix possible infinite loop in `useCheckoutOrderDetails`
This fixes a possible infinite loop in `useCheckoutOrderDetails` when
the redux store contains the details of another checkout order than the
one the hook is requested to fetch, it will try infinitely to reset the
details state because the reset action that is being used does not
clear that particular state. To fix this, we use the `resetCheckout` to
clear all checkout data instead.