fix: fix UserTraits type in analytics
This fixes the `UserTraits` type in analytics to allow the user to
specify properties that lives in either one of the types that
compose the union. For example, `countryCode` only exists in `User`
type but I cannot create a new variable typed `UserTraits` that
contains the `countryCode` property with the current type. This change
will allow that