chore: use lerna to create github releases
This removes the `CreateRelease` workflow in favour of using the
`--create-release` option of lerna version command which was recently
added. This might make the process more reliable since the
`CreateRelease` workflow sometimes was not triggered, even though the
tags were pushed with the correct personal access token.
Note that this will use the tag name as the name of the release in
github which differs from the current title format which prepended
`Release` before the tag name.