feat(client): add checkoutSessions clients
This adds the following changes:

- Adds new clients to manipulate the `checkoutSessions` resource;
- Adds new `SizeOrder` value to `ProductVariantAttributeType` enum;
- Adds missing extensions to linting and format scripts in
`package.json`;
- Fixes metadata type in `CheckoutOrder` to accept only strings and not
other values as the service does not allow other types;
- Fixes some properties in `CheckoutOrderItem` type that were using
types that were not correct.

This fixes some problems with jsdoc and types that were surfaced after
code review. Some properties of the CheckoutOrderItem type were changed
to correctly reflect what the endpoints return as well.