fix: buildQueryStringFromObject query construction
This fixes the way the query string is being constructed in
`buildQueryStringFromObject` to apply url encoding to the query values
to avoid being incorrectly interpreted by the browser. For example,
if a string contains a `&` character, it would not be url escaped
and if it was added to the browser address bar it would incorrectly
parse the query string generated by this method.