fix: fix tracking of uniqueViewId and previousUniqueViewId in analytics
This fixes a problem where if two or more consecutive calls
to analytics.page are added before the microtask queue has an
opportunity to run, the same values for the `uniqueViewId`
and `previousUniqueViewId` parameters for both the Omnitracking
events that would be generated. While this works, this is a temporary
fix since there are other values that suffer from the same problem,
for example, the consent and user values can be affected since they
are globally shared. A deeper change to analytics would be needed
to tackle this but as this is not a very common case, we will leave it
like that for now.